+++ src/main/java/mod/gottsch/forge/treasure2/core/block/DeferredWitherTreeGeneratorBlock.java
@@ -1,0 +1,66 @@
+/*
+ * This file is part of  Treasure2.
+ * Copyright (c) 2023 Mark Gottschling (gottsch)
+ *
+ * All rights reserved.
+ *
+ * Treasure2 is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * Treasure2 is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with Treasure2.  If not, see <http://www.gnu.org/licenses/lgpl>.
+ */
+package mod.gottsch.forge.treasure2.core.block;
+
+import mod.gottsch.forge.treasure2.Treasure;
+import mod.gottsch.forge.treasure2.core.block.entity.DeferredRandomVanillaSpawnerBlockEntity;
+import mod.gottsch.forge.treasure2.core.block.entity.DeferredWitherTreeGeneratorBlockEntity;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.EntityBlock;
+import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraft.world.level.block.entity.BlockEntityTicker;
+import net.minecraft.world.level.block.entity.BlockEntityType;
+import net.minecraft.world.level.block.state.BlockState;
+import org.jetbrains.annotations.Nullable;
+
+/**
+ *
+ * @author Mark Gottschling on Oct 12, 2023
+ *
+ */
+public class DeferredWitherTreeGeneratorBlock extends Block implements EntityBlock {
+
+    public DeferredWitherTreeGeneratorBlock(Properties properties) {
+        super(properties);
+    }
+
+    @Nullable
+    @Override
+    public BlockEntity newBlockEntity(BlockPos pos, BlockState state) {
+        Treasure.LOGGER.debug("created BE");
+        DeferredWitherTreeGeneratorBlockEntity blockEntity = new DeferredWitherTreeGeneratorBlockEntity(pos, state);
+        return (BlockEntity) blockEntity;
+    }
+
+    @javax.annotation.Nullable
+    @Override
+    public <T extends BlockEntity> BlockEntityTicker<T> getTicker(Level level, BlockState state, BlockEntityType<T> type) {
+        if (!level.isClientSide()) {
+            return (lvl, pos, blockState, t) -> {
+                if (t instanceof DeferredWitherTreeGeneratorBlockEntity entity) { // test and cast
+                    entity.tickServer();
+                }
+            };
+        }
+        return null;
+    }
+}