+++ src/main/java/mod/gottsch/forge/treasure2/core/block/entity/TreasureBlockEntities.java
@@ -119,6 +119,18 @@ public class TreasureBlockEntities {
 					TreasureBlocks.GRAVESTONE3_SPAWNER_OBSIDIAN.get()
 				).build(null));
 	
+	public static final RegistryObject<BlockEntityType<DeferredRandomVanillaSpawnerBlockEntity>> DEFERRED_RANDOM_VANILLA_SPAWNER_ENTITY_TYPE =
+			Registration.BLOCK_ENTITIES.register("deferred_random_vanilla_spawner",
+					() -> BlockEntityType.Builder.of(DeferredRandomVanillaSpawnerBlockEntity::new,
+							TreasureBlocks.DEFERRED_RANDOM_VANILLA_SPAWNER.get())
+							.build(null));
+
+	public static final RegistryObject<BlockEntityType<DeferredWitherTreeGeneratorBlockEntity>> DEFERRED_WITHER_TREE_GENERATOR_ENTITY_TYPE =
+			Registration.BLOCK_ENTITIES.register("deferred_wither_tree_generator",
+					() -> BlockEntityType.Builder.of(DeferredWitherTreeGeneratorBlockEntity::new,
+									TreasureBlocks.DEFERRED_WITHER_TREE_GENERATOR.get())
+							.build(null));
+	
 	public static final RegistryObject<BlockEntityType<MistEmitterBlockEntity>> MIST_EMITTER_BLOCK_ENTITY_TYPE = 
 			Registration.BLOCK_ENTITIES.register("mist_emitter", 
 				() -> BlockEntityType.Builder.of(MistEmitterBlockEntity::new,