+++ src/main/java/mod/gottsch/forge/treasure2/core/generator/template/ITemplateGenerator.java
@@ -29,7 +29,6 @@ import mod.gottsch.forge.gottschcore.world.gen.structure.PlacementSettings;
 import mod.gottsch.forge.treasure2.core.generator.GeneratorResult;
 import mod.gottsch.forge.treasure2.core.generator.IGeneratorResult;
 import mod.gottsch.forge.treasure2.core.generator.TemplateGeneratorData;
-import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockState;
 
@@ -58,28 +57,30 @@ public interface ITemplateGenerator<RESULT extends IGeneratorResult<?>> {
 	 * @param placement
 	 * @return
 	 */
-	public static ICoords alignEntranceToCoords(ICoords coords, ICoords entranceCoords, BlockPos size, PlacementSettings placement) {
-		ICoords startCoords = null;
-		// NOTE work with rotations only for now
-		
+	public static ICoords alignEntranceToCoords(ICoords coords, ICoords entranceCoords) {
+//		ICoords startCoords = null;
+		// NOTE this alignment works differently than  surface structure align - that is taking in the SIZE
+		// of the rotated structure. This method takes in an offset coords of the rotated structure
+
+		// NOTE this method is using a offset from the spawnpoint. so the rotation doesn't matter,
+		// it will always be a subtraction from the spawn coords. ie .add(-c) or .subtract(c)
+
+		// TODO this line is wrong - it only takes into account 1 rotation (90cw)
 		// first offset coords by entrance
-		startCoords = coords.add(-entranceCoords.getX(), 0, -entranceCoords.getZ());
-		
+		return coords.add(-entranceCoords.getX(), 0, -entranceCoords.getZ());
+
 		// make adjustments for the rotation. REMEMBER that pits are 2x2
-		switch (placement.getRotation()) {
-		case CLOCKWISE_90:
-			startCoords = startCoords.add(1, 0, 0);
-			break;
-		case CLOCKWISE_180:
-			startCoords = startCoords.add(1, 0, 1);
-			break;
-		case COUNTERCLOCKWISE_90:
-			startCoords = startCoords.add(0, 0, 1);
-			break;
-		default:
-			break;
-		}
-		return startCoords;
+//		return switch (placement.getRotation()) {
+//		case CLOCKWISE_90 ->
+//				coords.add(entranceCoords.getX()-1, 0, -(entranceCoords.getZ()-1));
+////			startCoords = startCoords.add(1, 0, 0);
+//		case CLOCKWISE_180 ->
+//				coords.add(entranceCoords.getX()-1, 0, entranceCoords.getZ()-1);
+////			startCoords = startCoords.add(1, 0, 1);
+//		case COUNTERCLOCKWISE_90 -> coords.add(-(entranceCoords.getX()-1), 0, entranceCoords.getZ()-1);
+////			startCoords = startCoords.add(0, 0, 1);
+//		default -> coords;
+//		};
 	}
 	
 	public Block getNullBlock();